Fix range of detected shortcut for quick switch.
To quickly switch to the desired tab, there is a shortcut ctrl-0 to
ctrl-9, where ctrl-0 will switch to tab 10, ctrl-1 to tab 1 and so on.
There is a range in the code, which detect which tab to pick up to jump.
Range was to wide, this commit is fixing that.