Fixed rename behaviour.
Renaming tab was broken by the fact, that renaming process provides
custom key_press callback just for it. After renaming was finished, all
the keypress shortcuts was gone, since it was done for main package.

The solution was to move the rename key press callback into tab scope,
so it wont affect main one.